+++ packages/ember/src/ember/entity-name/camelize.ts
