+++ packages/ember/src/ember/entity-name/classify.ts
