+++ packages/ember/src/ember/entity-name/double-colonize.ts
