+++ packages/ember/tests/ember/entity-name/camelize/edge-case-entity-name.test.ts
@@ -1,8 +1,8 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { camelize } from '../../../src/index.js';
+import { camelize } from '../../../../src/index.js';
 
-test('entity-name | camelize > edge case (entity name)', function () {
+test('utils | ember | entity-name | camelize > edge case (entity name)', function () {
   assert.strictEqual(camelize('tracks'), 'tracks');
   assert.strictEqual(camelize('navigation-menu'), 'navigationMenu');
   assert.strictEqual(camelize('ui/page'), 'ui/page');
