+++ packages/ember/tests/ember/entity-name/classify/base-case.test.ts
@@ -1,8 +1,8 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { classify } from '../../../src/index.js';
+import { classify } from '../../../../src/index.js';
 
-test('entity-name | classify > base case', function () {
+test('utils | ember | entity-name | classify > base case', function () {
   assert.strictEqual(classify('innerHTML'), 'InnerHTML');
   assert.strictEqual(classify('action_name'), 'ActionName');
   assert.strictEqual(classify('css-class-name'), 'CssClassName');
