+++ packages/ember/tests/ember/entity-name/classify/edge-case-periods.test.ts
@@ -1,8 +1,8 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { classify } from '../../../src/index.js';
+import { classify } from '../../../../src/index.js';
 
-test('entity-name | classify > edge case (periods)', function () {
+test('utils | ember | entity-name | classify > edge case (periods)', function () {
   assert.strictEqual(classify('aa.bb.cc'), 'Aa.Bb.Cc');
   assert.strictEqual(classify('aa.b/b.cc'), 'Aa.BB.Cc');
 });
