+++ packages/ember/tests/ember/entity-name/classify/edge-case-spaces.test.ts
@@ -1,8 +1,8 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { classify } from '../../../src/index.js';
+import { classify } from '../../../../src/index.js';
 
-test('entity-name | classify > edge case (spaces)', function () {
+test('utils | ember | entity-name | classify > edge case (spaces)', function () {
   assert.strictEqual(classify('my favorite items'), 'MyFavoriteItems');
   assert.strictEqual(classify('My favorite items'), 'MyFavoriteItems');
 });
