+++ packages/ember/tests/ember/entity-name/double-colonize/base-case.test.ts
@@ -1,8 +1,8 @@
 import { assert, test } from '@codemod-utils/tests';
 
-import { doubleColonize } from '../../../src/index.js';
+import { doubleColonize } from '../../../../src/index.js';
 
-test('entity-name | double-colonize > base case', function () {
+test('utils | ember | entity-name | double-colonize > base case', function () {
   assert.strictEqual(doubleColonize('innerHTML'), 'InnerHTML');
   assert.strictEqual(doubleColonize('action_name'), 'ActionName');
   assert.strictEqual(doubleColonize('css-class-name'), 'CssClassName');
