--- packages/ember/tests/entity-name/classify/edge-case-empty-string.test.ts
@@ -1,7 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { classify } from '../../../src/index.js';
-
-test('entity-name | classify > edge case (empty string)', function () {
-  assert.strictEqual(classify(''), '');
-});
