--- packages/ember/tests/entity-name/double-colonize/edge-case-empty-string.test.ts
@@ -1,7 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { doubleColonize } from '../../../src/index.js';
-
-test('entity-name | double-colonize > edge case (empty string)', function () {
-  assert.strictEqual(doubleColonize(''), '');
-});
